+++ src/components/ForgotPassword/index.ts
