+++ src/pages/ForgotPasswordPage/index.ts
@@ -1,0 +1,1 @@
+export { ForgotPasswordPage } from './ForgotPasswordPage'
