+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,16 +1,16 @@
 import { IonContent, IonPage } from '@ionic/react'
 import { ReactElement } from 'react'
-import { Login as LoginComponent } from '../../components/Login'
+import { Login } from '../../components/Login'
 
 /* Due to a bug with testing this page component has been separated from the
    implementation and instead just wraps the implementation in an ionic page.
    https://github.com/ionic-team/stencil/issues/3434
   */
-export function Login(): ReactElement {
+export function LoginPage(): ReactElement {
   return (
     <IonPage>
       <IonContent fullscreen>
-        <LoginComponent />
+        <Login />
       </IonContent>
     </IonPage>
   )
